+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Changelog
 * Change 'Unassigned' genotype to 'Indeterminate' genotype.
 * Add new method :meth:`api.core.get_variant_synonyms`.
 * Update :meth:`api.core.list_variants` method to accept multiple star alleles.
+* Update :command:`predict-alleles` command to support multiallelic variants.
 
 0.5.0 (2021-10-02)
 ------------------